Add supports for multiple themes of APlayer
This commit adds the feature that multiple (light and dark) themes are
added to APlayer source code and that the theme is switched based on the
theme (light and dark) of the blog.

metowolf/MetingJS was dependent on DIYgod/APlayer and was used in the
old blog theme. The current APlayer hasn't been updated for long and it
doesn't support other themes. Then the committer forked the repository
to LittleYe233/APlayer and tried to support the feature, code details
and the changelog seen in it.

And considering that requests to cdn.jsdelivr.net from clients in China
Mainland are unstable, all resource requests to the website are saved to
the repository. Specifically, they are APlayer.min.css and Meting.min.js
which are maintained by LittleYe233/APlayer and metowolf/MetingJS res-
pectively.

Some files also have changes in code styles by ESLint.