Add "enabled" option to comment rendering
This commit mainly reconstructed the trigger option of rendering the
comments, including those below the body of an article and "Hot Recom-
mended" and "Latest Comments" cards at the top of the homepage, which
can be affected by the type of the active commenting extension.

There was no trigger option like "enable" in `hexo-theme-fluid` in the
original configurations and the ways to disable commenting cards can
be inconvenient (i.e. commenting all configuration lines of "comment"
block, which may be annoying in some editors) or users would received
warnings from the consoles. Therefore, an "enabled" option was added
to make all the rest configurations related to comments inactive when
its value is "false" (a boolean value). More information will be showed
in the configuration manual.

Please note that this type of design may be applied to other blocks in
_config.yml.