Add `show_mode` configuration field to widgets
This new field has two available values now: `pages_and_posts` and
`default` whose meanings are written in `_config.yml`.

In the previous versions *all* the widgets were shown in the left column
of pages with a layout of pages and posts, no matter what the values of
the `position` field of these widgets are. It brought an issue that the
table of contents might be rolled up before the page reached the end of
a post, for the `card` elements containing these widgets have special
CSS styles.

Now if a widget in `_config.yml` has the `show_mode` field with a value
`pages_and_posts`, it will be shown in the left column when it comes to
a page or a post like what in the old version. But if it is set to
`default`, it won't appear. The default setting is that you can only see
the table of contents on a page or a post.