Try to add support for KaTeX but failed
We tried to add support for KaTeX more than MathJax but failed because
the current renderer (`hexo-renderer-inferno`) may have some bugs on
Markdown parsing, for the lines as below:
```md
$$
a+b=c
$$
```
shouldn't be rendered to: `$$<br>a+b+c<br>$$`, but `$$a+b+c$$` instead.
Now it's recommended to use MathJax only. Configurations of KaTeX are
commented now.